+++ core/src/test/java/apoc/schema/SchemasTest.java
@@ -682,58 +682,6 @@ public class SchemasTest {
         });
     }
 
-    @Test
-    public void testCompareIndexesAndConstraints() {
-        db.executeTransactionally("CREATE INDEX FOR (n:Person) ON (n.surname)");
-        db.executeTransactionally("CREATE CONSTRAINT FOR (n:Person) REQUIRE (n.address) IS UNIQUE");
-        db.executeTransactionally("CREATE INDEX FOR (n:Movie) ON (n.name)");
-        db.executeTransactionally("CALL db.index.fulltext.createNodeIndex('fullSecondIdx', ['Person', 'Another'], ['weightProp'])");
-
-        testResult(db, "CALL apoc.schema.node.compareIndexesAndConstraints()", (res) -> {
-            Map<String, Object> row = res.next();
-            assertEquals(Map.of(":[Another, Person],(weightProp)", List.of("weightProp")), row.get("onlyIdxProps"));
-            assertEquals("Another", row.get("label"));
-            assertEquals(Collections.emptyMap(), row.get("onlyConstraintsProps"));
-            assertEquals(Collections.emptyList(), row.get("commonProps"));
-            row = res.next();
-            assertEquals(Map.of(":Movie(name)", List.of("name")), row.get("onlyIdxProps"));
-            assertEquals("Movie", row.get("label"));
-            assertEquals(Collections.emptyMap(), row.get("onlyConstraintsProps"));
-            assertEquals(Collections.emptyList(), row.get("commonProps"));
-            row = res.next();
-            assertEquals(Map.of(":[Another, Person],(weightProp)", List.of("weightProp"), ":Person(surname)", List.of("surname")), row.get("onlyIdxProps"));
-            assertEquals("Person", row.get("label"));
-            assertEquals(Collections.emptyMap(), row.get("onlyConstraintsProps"));
-            assertEquals(List.of("address"), row.get("commonProps"));
-            assertFalse(res.hasNext());
-        });
-    }
-    
-    @Test
-    public void testCompareIndexesAndConstraintsRel() {
-        db.executeTransactionally("CALL db.index.fulltext.createRelationshipIndex('fullIdxRel', ['TYPE_1', 'TYPE_2'], ['alpha', 'beta'])");
-        db.executeTransactionally("CREATE INDEX rel_index_name FOR ()-[r:KNOWS]-() ON (r.since)");
-
-        testResult(db, "CALL apoc.schema.relationship.compareIndexesAndConstraints()", (res) -> {
-            Map<String, Object> row = res.next();
-            assertEquals("KNOWS", row.get("type"));
-            assertEquals(Map.of(":KNOWS(since)", List.of("since")), row.get("onlyIdxProps"));
-            assertEquals(Collections.emptyMap(), row.get("onlyConstraintsProps"));
-            assertEquals(Collections.emptyList(), row.get("commonProps"));
-            row = res.next();
-            assertEquals("TYPE_1", row.get("type"));
-            assertEquals(Map.of(":[TYPE_1, TYPE_2],(alpha,beta)", List.of("alpha", "beta")), row.get("onlyIdxProps"));
-            assertEquals(Collections.emptyMap(), row.get("onlyConstraintsProps"));
-            assertEquals(Collections.emptyList(), row.get("commonProps"));
-            row = res.next();
-            assertEquals(Map.of(":[TYPE_1, TYPE_2],(alpha,beta)", List.of("alpha", "beta")), row.get("onlyIdxProps"));
-            assertEquals("TYPE_2", row.get("type"));
-            assertEquals(Collections.emptyMap(), row.get("onlyConstraintsProps"));
-            assertEquals(Collections.emptyList(), row.get("commonProps"));
-            assertFalse(res.hasNext());
-        });
-    }
-
     private void dropSchema()
     {
         try(Transaction tx = db.beginTx()) {
