+++ full/src/main/java/apoc/result/CompareIdxToCons.java
