+++ full/src/main/java/apoc/result/CompareIdxToConsNodes.java
